fix: moved keyDerivation.ts to utils
